centered the clock div and added an h1 tag
