+++ src/components/Auto/Clock/Clock.jsx
@@ -1,4 +1,5 @@
 import React, { Component } from 'react';
+import './clock.css';
 
 class Clock extends Component {
   constructor(props) {
@@ -19,12 +20,10 @@ class Clock extends Component {
     setInterval(() => this.getTimeUntil(this.props.deadline), 1000);
   }
 
-
-
   getTimeUntil(deadline) {
     const time = Date.parse(deadline) - Date.parse(new Date());
     const seconds = Math.floor((time / 1000) % 60);
-    const minutes = Math.floor(time / 1000 * 60);
+    const minutes = Math.floor(time / (1000 * 60) % 60);
     const hours = Math.floor((time / (1000 * 60 * 60)) % 24);
     const days = Math.floor(time / (1000 * 60 * 60 * 24));
 
@@ -37,11 +36,11 @@ class Clock extends Component {
 
   render() {
     return (
-      <div>
-        <div className="Clock-days">Days:{this.state.days}</div>
-        <div className="Clock-hours">Hours:{this.leadingZero(this.state.hours)}</div>
-        <div className="Clock-minutes">Minutes:{this.leadingZero(this.state.minutes)}</div>
-        <div className="Clock-seconds">Seconds:{this.leadingZero(this.state.seconds)}</div>
+      <div className='Clock-container'>
+        <div className='Clock-days'>Days:{this.state.days}</div>
+        <div className='Clock-hours'>Hours:{this.leadingZero(this.state.hours)}</div>
+        <div className='Clock-minutes'>Minutes:{this.leadingZero(this.state.minutes)}</div>
+        <div className='Clock-seconds'>Seconds:{this.leadingZero(this.state.seconds)}</div>
       </div>
     );
   }
