+++ src/components/Auto/Clock/clock.css
@@ -1,11 +1,10 @@
 .App {
     font-size: 15px;
-    margin-top: 150px;
     height: 330px;
     width: 350px;
-    margin-left: 30px;
     box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2);
     transition: 0.3s;
+    margin: 0 auto;
 }
 
 .App-title {
@@ -33,4 +32,5 @@
 
 .paragraph {
     font-weight: bold;
+    margin: 3px;
 }
