+++ src/pages/auto.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import CardCheck from '../components/Auto/CardCheck/CardCheck';
-import ClockApp from '../components/Auto/Offer/ClockApp';
+import ClockApp from '../components/Auto/Clock/ClockApp';
 
 const Auto = () => (
   <div>
